Moved MetricsConfiguration to a common component
