+++ components/org.wso2.carbon.metrics.common/src/test/java/org/wso2/carbon/metrics/common/Utils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.carbon.metrics.impl;
+package org.wso2.carbon.metrics.common;
 
 import java.net.URL;
