+++ components/org.wso2.carbon.metrics.common/src/test/resources/metrics.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<!--
+    ~ Copyright 2015 WSO2 Inc. (http://wso2.org)
+    ~
+    ~ Licensed under the Apache License, Version 2.0 (the "License");
+    ~ you may not use this file except in compliance with the License.
+    ~ You may obtain a copy of the License at
+    ~
+    ~ http://www.apache.org/licenses/LICENSE-2.0
+    ~
+    ~ Unless required by applicable law or agreed to in writing, software
+    ~ distributed under the License is distributed on an "AS IS" BASIS,
+    ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    ~ See the License for the specific language governing permissions and
+    ~ limitations under the License.
+-->
+<Metrics xmlns="http://wso2.org/projects/carbon/metrics.xml">
+
+    <Level>ALL</Level>
+
+    <Reporting>
+        <CSV>
+            <Enabled>false</Enabled>
+            <Location>${carbon.home}/repository/logs/metrics/</Location>
+            <PollingPeriod>60</PollingPeriod>
+        </CSV>
+    </Reporting>
+</Metrics>
